+++ src/array/clear.js
@@ -1,0 +1,10 @@
+'use strict'
+
+/**
+ * Clear array
+ *
+ * @param {Array} array
+ */
+export default function clear(array) {
+    array.length = 0
+}