+++ src/array/index.js
@@ -1,7 +1,18 @@
 'use strict'
 
 import equals from './equals'
+import clear from './clear'
+import intersect, {intersectAll} from './intersect'
+import difference from './difference'
+import symmetricalDifference from './symmetricalDifference'
+import random from './random'
 
 export {
     equals,
+    clear,
+    intersect,
+    intersectAll,
+    difference,
+    symmetricalDifference,
+    random,
 }