+++ src/array/random.js
@@ -1,0 +1,11 @@
+'use strict'
+
+/**
+ * Random function returns random item from array
+ *
+ * @param {Array} array
+ * @returns {unknown}
+ */
+export default function random(array) {
+    return array[(Math.floor(Math.random() * (array.length)))]
+}